lookwhosstreaming/requirements.txt
amqp==2.6.0
asgiref==3.2.7
autopep8==1.5.2
billiard==3.6.3.0
boto3==1.12.46
botocore==1.15.46
celery==4.4.6
certifi==2020.6.20
chardet==3.0.4
cycler==0.10.0
dj-database-url==0.5.0
Django==3.0.5
django-celery-beat==2.0.0
django-heroku==0.3.1
django-storages==1.9.1
django-timezone-field==4.0
django-tinymce==3.0.2
docutils==0.15.2
future==0.18.2
gunicorn==20.0.4
idna==2.10
instagram-private-api @ git+https://git@github.com/ping/instagram_private_api.git@1d70e99bc11591161b0cd71cff3f0c08cd04b34f
jmespath==0.9.5
kiwisolver==1.2.0
kombu==4.6.11
matplotlib==3.2.2
numpy==1.19.0
Pillow==7.1.2
psycopg2==2.8.5
pycodestyle==2.5.0
pyparsing==2.4.7
python-crontab==2.5.1
python-dateutil==2.8.1
python-dotenv==0.13.0
pytz==2019.3
redis==3.5.3
requests==2.24.0
s3transfer==0.3.3
six==1.14.0
sqlparse==0.3.1
urllib3==1.25.9
vine==1.3.0
whitenoise==5.0.1
tweepy @ git+https://github.com/tweepy/tweepy.git
